Upgrade script sending images to a Telegram channel
The script now can be run from terminal, the send_to_channel.py
program allows to sending images to a Telegram channel automatically.
It mixed photos in the specifyed directory and sends all of them constantly
in the given interval of time. By default it is 4 hours. You can also specify
the directory of the images. By default it is "images"